requirements.txt
Flask-JWT
Flask-RESTful
Flask-SQLAlchemy
Flask-JWT-Extended